--- product-service/src/main/java/com/monkey/productservice/application/service/v3/ProductServiceApiV3.java
@@ -1,7 +1,0 @@
-package com.monkey.productservice.application.service.v3;
-
-import java.util.UUID;
-
-public interface ProductServiceApiV3 {
-    void increaseStock(UUID productId, int quantity);
-}
